test/index.html

<!doctype html>

<meta charset=utf-8>

<title>Category: Describe interface</title>

<link rel=match href="${2:URL of match}">
<link rel=help  href=//path/to/specification>

<script src=testharness.js></script>
<script src=testharnessreport.js></script>

<!-- Testing Format http://web-platform-tests.org/appendix/test-templates.html -->

<script>
//console.log   ('logging', 2+2)
//console.warn  ('warn', 2+2)
//console.error ('error', 2+2)
/*
  setup ( {} )

  // Example Group

  test (test => {
    // Test case

    assert_true (true, 'Test stub works!')
  }, 'Descriptive test name')

*/
</script>


<style>
    ${3:Test CSS}
</style>


${4:Test content}


<output id=log></output>
